pyproject.toml
[tool.poetry]
name = "flip-gpt-labs"
version = "0.1.0"
description = "Connect ChatGPT + ElevenLabs + OBS to chat live with video game characters"
authors = ["Angela Kochoska <a.kochoska@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.6.1"
elevenlabs = "^0.2.27"
sounddevice = "^0.4.6"
PyAudio = "^0.2.14"
langchain = "^0.0.352"
ipykernel = "^6.28.0"
jupyterlab = "^4.0.9"
langchainhub = "^0.1.14"
google-search-results = "^2.4.2"
python-dotenv = "^1.0.0"
kivy = "^2.2.1"
soundfile = "^0.12.1"
kivymd = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
